GUI to database connection tested
Check delete_me.py if you'd like to see it in action. Still haven't
figured out how to return the selection to the calling function, but
hopefully the interface team will solve that issue if they haven't
already.

Adjusted Print statements to Print functions (required by python 3.4)

added a config file to hold the database name so that we can create a
test database if needed. Feel free to add other variables if you feel
the need.

Changed search_database so that it can return entire columns. This will
increase the burden on anyone calling the function, but will enable the
creation of drop-down menus in the GUI.

Will work on databases when I have time. Too late right now.