Various small updates
config can now generate a list of all tables in the database to be
iterated over

csv files created for every table (no attributes defined in them yet,
will do tomorrow)

csv_to_sql simplified to only need one variable

Not sure if characters should be put into a database or how to
add/update them if they are.

I'll try to finish converting class objects to tables tomorrow (to the
extent to which that is possible without feedback) so that the csv's can
start getting filled.