minor bug fix
Forgot to test before commiting :/

I think we'll need to go through the CSV's and delete the ending comma's
or re-write the code for populating the tables. Right now, its reading
that last comma as an entry and rejecting the row for having too many
elements. Or at least, I think that's why its complaining when I try to
populate the database.