minor update to csv_to_sql
Fixed the problem with the extra element being generated by the trailing
comma, but it still won't load the majority of the entries. I'm not sure
why. Can someone else look at it and see if they can figure it out? Run
test_db_pop_and_search if you want to see it failing.